Added mouse controls, click on text boxes and buttons
I have added mouse controls allowing you to click on text boxes to select which one you are typing in, and I added a class that automatically draws a rectangle and words on it that can be clicked on. I also added the file RauchSprite.h and the shaders that go with it in preperation of adding a pause and play sprite. Then I modified the Shader.h file to make sure the ifstream exception functions are only compiled on Windows, and changed the Text class in RauchGUI.h to not scale with the screen. Finally, I recompiled the included x64 Windows .exe.