Added pause, shuffle, and playlists
I added the ability to load playlists of audio files. I also added a TextTable struct to RauchGUI.h to display the playlist on the screen. I also added image files for pause, play, shuffle, and looping, so they could be rendered to the screen to show if the currently active song is paused or looping, or if shuffle is enabled. I also wrote a brief README.txt file, and finally I recompiled the included Windows x64 .exe file.