Cleaned up code and fixed some problems
Fixed Saving (https://github.com/Huntergrauch/The-Rauch-Music-Player/issues/1) to first check if a folder for the Playlists has been created, and if not create one.

Added comments to most functions to explain what they do.

Fixed DeleteFont() in RauchGUI.h, which caused a segfault on Linux.

Removed RauchInput.h, moved the code in it for RauchGUI.h into RauchGUI.h, the rest I moved into Program.cpp.

Added a LICENSE.txt to specify this program's license.

Changed the executable icon and tray icon.