Fixed possible memory leaks and added pausing on click
Added a Delete() funciton to all classes with opengl VBOs and VAOs.

Added a method to check if a Sprite has been clicked on int RauchSprite.h. Used this method to add the ability to click on the pause, shuffle and looping sprites to pause shuffle or loop the music.

Finally, I recompiled the x64 windows executable.