Renamed Shaders and Changed paths
Renamed the shaders to shorter names.

Changed all the paths to be easier to change.

Added an option to define LINUX_INSTALL to change all the paths to absolute paths that point to where the program would be installed in on the Linux filesystem, and added an option when building with CMake to enable that define.

Finally I started a system for agument parsing.