Adjusted path system for Windows
Adjusted the path system when on Windows to use the Windows function GetModuleFileName() to find the path to the executable rather than relying on the period for relative paths such as "./rmpResources".

Also added the the Visual Studio project and solution files to .gitignore as those are meant to be built via CMake.